Combining spring-cloud-azure-data-cosmos-sample and spring-cloud-azure-data-cosmos-aks-sample sample sample.
