--- cosmos/spring-cloud-azure-starter-data-cosmos/spring-cloud-azure-data-cosmos-sample/src/main/java/com/azure/spring/sample/cosmos/UserRepository.java
@@ -1,14 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-package com.azure.spring.sample.cosmos;
-
-import com.azure.spring.data.cosmos.repository.ReactiveCosmosRepository;
-import org.springframework.stereotype.Repository;
-import reactor.core.publisher.Flux;
-
-@Repository
-public interface UserRepository extends ReactiveCosmosRepository<User, String> {
-
-    Flux<User> findByFirstName(String firstName);
-}
