--- cosmos/spring-cloud-azure-starter-data-cosmos/spring-cloud-azure-data-cosmos-sample/src/main/resources/application.yml
@@ -1,7 +1,0 @@
-spring:
-  cloud:
-    azure:
-      cosmos:
-        endpoint: ${AZURE_COSMOS_ENDPOINT}
-        database: ${COSMOS_DATABASE}
-
