--- cosmos/spring-cloud-azure-starter-data-cosmos/spring-cloud-azure-data-cosmos-sample/terraform/setup_env.sh
@@ -1,5 +1,0 @@
-export AZURE_COSMOS_ENDPOINT=$(terraform -chdir=./terraform output -raw azure_cosmos_endpoint)
-export COSMOS_DATABASE=$(terraform -chdir=./terraform output -raw cosmos_database_name)
-
-echo AZURE_COSMOS_ENDPOINT=$AZURE_COSMOS_ENDPOINT
-echo COSMOS_DATABASE=$COSMOS_DATABASE
